update Emploeye ID to UUID

+++ src/main/java/by/antohakon/vetclinicemployes/service/EmployeServiceImpl.java
@@ -13,6 +13,8 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import java.util.UUID;
+
 @Service
 @RequiredArgsConstructor
 @Slf4j
@@ -27,7 +29,7 @@ public class EmployeServiceImpl implements EmployeService {
         log.info("getAllEmployes");
         return employeRepository.findAll(pageable)
                 .map(employe -> EmployeDto.builder()
-                        .id(employe.getId())
+                        .doctorId(employe.getEmployeeId())
                         .lastName(employe.getLastName())
                         .firstName(employe.getFirstName())
                         .role(employe.getRole())
@@ -35,19 +37,21 @@ public class EmployeServiceImpl implements EmployeService {
     }
 
     @Override
-    public EmployeDto getEmployeById(Long id) {
+    public EmployeDto getEmployeById(UUID id) {
 
         log.info("getEmployeById");
-        Employe findEmploye = employeRepository.findById(id)
-                .orElseThrow(() -> new EmployeNotFoundException("Emplye not found with id " + id));
+        Employe findEmploye = employeRepository.findByEmployeeId(id);
+        if (findEmploye == null) {
+            throw new  EmployeNotFoundException("Emplye not found with id " + id);
+        }
 
-        return employeMapper.toDto(findEmploye);
-//        return EmployeDto.builder() - это если снова перестанет работать mapstruct
-//                .id(findEmploye.getId())
-//                .lastName(findEmploye.getLastName())
-//                .firstName(findEmploye.getFirstName())
-//                .role(findEmploye.getRole())
-//                .build();
+      //  return employeMapper.toDto(findEmploye);
+        return EmployeDto.builder()
+                .doctorId(findEmploye.getEmployeeId())
+                .lastName(findEmploye.getLastName())
+                .firstName(findEmploye.getFirstName())
+                .role(findEmploye.getRole())
+                .build();
     }
 
     @Override
@@ -58,47 +62,59 @@ public class EmployeServiceImpl implements EmployeService {
             throw new EmployeDuplicationException("Employee already exists with last name " + employe.lastName());
         }
 
-         Employe newEmployee = employeMapper.toEntity(employe);
-//        Employe newEmployee = Employe.builder()  - это если снова перестанет работать mapstruct
-//                .lastName(employe.lastName())
-//                .firstName(employe.firstName())
-//                .role(employe.role())
-//                .build();
+        // Employe newEmployee = employeMapper.toEntity(employe);
+        Employe newEmployee = Employe.builder()
+                .employeeId(UUID.randomUUID())
+                .lastName(employe.lastName())
+                .firstName(employe.firstName())
+                .role(employe.role())
+                .build();
 
         employeRepository.save(newEmployee);
 
-      return employeMapper.toDto(newEmployee);
-//        return EmployeDto.builder() - это если снова перестанет работать mapstruct
-//                .id(newEmployee.getId())
-//                .lastName(newEmployee.getLastName())
-//                .firstName(newEmployee.getFirstName())
-//                .role(newEmployee.getRole())
-//                .build();
+  //    return employeMapper.toDto(newEmployee);
+        return EmployeDto.builder()
+                .doctorId(newEmployee.getEmployeeId())
+                .lastName(newEmployee.getLastName())
+                .firstName(newEmployee.getFirstName())
+                .role(newEmployee.getRole())
+                .build();
 
     }
 
     @Override
-    public EmployeDto updateEmploye(Long id, CreateEmployeDto newEmploye) {
+    public EmployeDto updateEmploye(UUID id, CreateEmployeDto newEmploye) {
 
         log.info("updateEmploye");
-        Employe findEmploye = employeRepository.findById(id)
-                .orElseThrow(() -> new EmployeNotFoundException("Emplye not found with id " + id));
+        Employe findEmploye = employeRepository.findByEmployeeId(id);
+        if (findEmploye == null) {
+           throw new EmployeNotFoundException("Emplye not found with id " + id);
+        }
 
         findEmploye.setLastName(newEmploye.lastName());
         findEmploye.setFirstName(newEmploye.firstName());
         findEmploye.setRole(newEmploye.role());
         employeRepository.save(findEmploye);
 
-        return employeMapper.toDto(findEmploye);
+       // return employeMapper.toDto(findEmploye);
+        EmployeDto employeDto = EmployeDto.builder()
+                .doctorId(findEmploye.getEmployeeId())
+                .lastName(findEmploye.getLastName())
+                .firstName(findEmploye.getFirstName())
+                .role(findEmploye.getRole())
+                .build();
 
+        return employeDto;
     }
 
     @Override
-    public void deleteEmploye(Long id) {
+    public void deleteEmploye(UUID id) {
 
         log.info("deleteEmploye");
-        Employe findEmploye = employeRepository.findById(id)
-                .orElseThrow(() -> new EmployeNotFoundException("Emplye not found with id " + id));
+        Employe findEmploye = employeRepository.findByEmployeeId(id);
+        if (findEmploye == null) {
+           throw new EmployeNotFoundException("Emplye not found with id " + id);
+        }
 
         employeRepository.delete(findEmploye);
